Refactor `SimpleLinkedList` and unify `Empty` and `Node`.
Replaces `Empty` with `EmptyList` and updates `SimpleLinkedList` to use a sealed trait. Consolidates related logic for better type safety, readability, and maintainability, while retaining core functionality.
